requirements.txt
Django==6.0.1
djangorestframework==3.16.1
psycopg[binary]==3.3.2
gunicorn==24.0.0
python-dotenv==1.2.1
requests==2.32.3
python-dotenv==1.2.1
